Disable iOS target for now Move authentication code to JVM specific source set Remove `@Language` annotations since it does not work in `commonMain`
