+++ goplay/src/commonTest/kotlin/be/tapped/goplay/e2e/ProfileRepoTest.kt
@@ -1,8 +1,7 @@
 package be.tapped.goplay.e2e
 
 import be.tapped.goplay.Credentials
-import be.tapped.goplay.profile.HttpProfileRepo
-import be.tapped.goplay.profile.ProfileUserAttributeParser
+import be.tapped.goplay.profile.httpProfileRepo
 import io.kotest.assertions.arrow.core.shouldBeLeft
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.FreeSpec
@@ -10,7 +9,7 @@ import io.kotest.core.spec.style.FreeSpec
 internal class ProfileRepoTest : FreeSpec({
     "with valid credentials" - {
         val (username, password) = Credentials.default
-        val sut = HttpProfileRepo(ProfileUserAttributeParser())
+        val sut = httpProfileRepo
 
         "should be able to fetch tokens" {
             val token = sut.fetchTokens(username, password).shouldBeRight().token
@@ -28,7 +27,7 @@ internal class ProfileRepoTest : FreeSpec({
     }
 
     "with non valid credentials" - {
-        val sut = HttpProfileRepo(ProfileUserAttributeParser())
+        val sut = httpProfileRepo
 
         "it should not be able to fetch tokens" {
             sut.fetchTokens("john-doe", "password").shouldBeLeft()
